Fix downscaling of underheight fragments
If a line of fragments didn't fill the full height, then when
scaling down to fit it might still fit within the adjusted
height without scaling (which would not cause it to scale
down in X). This fixes by always scaling by the actual rendered
height, instead of the available area.