Add an 'embedded' label style
This has the label flush with the top surface of the base. The label is
cut into the base, but filled. This is for printing on multimaterial
printers.

This has included slightly redefining directions of cuts and what is expected
from a base generator. The base is returned such that the label is always at
the surface, and it's either cut in or extruded up depending on the style.

A side effect of this is that --depth actually affects debossed labels now.