Modern: Handle configurable depth
The standard (as of Rev 2.3) has varying label depth depending on
the height of the acetate window that you have inserted into the
box. 2.2mm is the default, by virtue of the fact that the default
(and direct STL downloads on printables) is configured using a
0.8 mm default (+0.2mm tolerance +1.2mm label).

Also, removed the tolerances for now as we understand the origin
of the changes. These may need to be added back, because the default
labels have zero tolerance above the gap they are trying to fit.